+++ dlib/optimization/find_max_factor_graph_nmplp.h
@@ -40,7 +40,7 @@ namespace dlib
                     - #(*this)(a,b) == value
             !*/
             {
-                const uint32 h = murmur_hash3_2(a,b)%(data.size()-scan_dist);
+                const uint32 h = (unsigned int)murmur_hash3_2((unsigned int)a,(unsigned int)b)%(data.size()-scan_dist);
 
                 const unsigned long empty_bucket = std::numeric_limits<unsigned long>::max();
 
@@ -84,7 +84,7 @@ namespace dlib
                 DLIB_ASSERT(a != b, "An invalid map_problem was given to find_max_factor_graph_nmplp()."
                             << "\nNode " << a << " is listed as being a neighbor with itself, which is illegal.");
 
-                uint32 h = murmur_hash3_2(a,b)%(data.size()-scan_dist);
+                uint32 h = murmur_hash3_2((unsigned int)a,(unsigned int)b)%(data.size()-scan_dist);
 
 
                 for (unsigned long i = 0; i < scan_dist; ++i)
